moving typing related objects into typing.py, fix #713
