+++ tests/test_utils.py
@@ -4,17 +4,8 @@ from typing import NewType, Union
 
 import pytest
 
-from pydantic.utils import (
-    ValueItems,
-    display_as_type,
-    import_string,
-    is_new_type,
-    lenient_issubclass,
-    make_dsn,
-    new_type_supertype,
-    truncate,
-    validate_email,
-)
+from pydantic.typing import display_as_type, is_new_type, lenient_issubclass, new_type_supertype
+from pydantic.utils import ValueItems, import_string, make_dsn, truncate, validate_email
 
 try:
     import email_validator
